# Add a time-based trigger to delete trace entries older than 30 days.