This will help keep the size of the trace table in check.

NOTE: an earlier iteration of this change used to accept the time horizon value (i.e., 30 days) as an option. However, since the database can be shared across multiple clients with potentially different options, there is no way to guarantee that the time horizon value will be respected across the clients. For this reason, we hard-code a reasonable value for the time horizon of 30 days, at least for now.

Performance neutral (at least when no deletions are happening).

Before:
BenchmarkStore/1-72  	     100	  10824509 ns/op
BenchmarkStore/10-72 	     100	  11441102 ns/op
BenchmarkStore/100-72         	      91	  13537496 ns/op

After:
BenchmarkStore/1-72  	     100	  10735265 ns/op
BenchmarkStore/10-72 	     100	  11530419 ns/op
BenchmarkStore/100-72         	      93	  14008470 ns/op

Other changes:
  * Delete the attributes table as it is currently unused.