@spetrovic77 2023-07-18
Note: I had to implement a two-trigger solution as suggested by @ghemawat offline. `PRAGMA foreign_keys=ON` doesn't work as spans are collected asynchronously, so a span may be inserted before its root span (i.e., its trace), which fails the foreign key enforcement.